+++ src/components/card/CardDouble.js
@@ -1,17 +1,13 @@
 import React from "react";
 import { styled } from "linaria/react";
 import { Link } from "gatsby";
-import pageLinksStyles from "./pagelinks.module.css";
 import Img from "gatsby-image";
-import EventDate from "./EventDate";
+import EventDate from "../EventDate";
+import CardContent from "./CardContent";
 
 const ImageCont = styled.div`
   margin: 0;
   position: relative;
-`;
-
-const CardContent = styled.div`
-  padding: 1rem 3px;
 `;
 
 const EventDateCont = styled.div`
@@ -30,11 +26,19 @@ const EventDateCont = styled.div`
 //   text-align: center;
 // `;
 
-const CardSingle = (props) => {
+const CardDouble = (props) => {
   let cardImage;
-  if (props.smallImage) {
-    console.log("smallImage", props.smallImage);
-    cardImage = <Img fixed={props.smallImage.childImageSharp.fixed} alt="" />;
+  if (props.largeImage) {
+    console.log("largeimage", props.largeImage);
+    cardImage = (
+      <Img
+        fluid={props.largeImage.childImageSharp.fluid}
+        alt=""
+        imgStyle={{ objectFit: "contain" }}
+        backgroundColor={true}
+        objectFit="contain"
+      />
+    );
   } else {
     cardImage = <img src="/img/default-image.jpg" alt="image" />;
   }
@@ -55,20 +59,16 @@ const CardSingle = (props) => {
           {cardImage}
         </ImageCont>
       </Link>
-      <CardContent>
-        <div className="content">
-          <h3>{props.title}</h3>
-          {props.displayLocation && <span>{props.location}</span>}
-          <p>{props.linkText}</p>
-          {props.showPageLink && (
-            <Link to={props.url} className={pageLinksStyles.btn}>
-              Find out more
-            </Link>
-          )}
-        </div>
-      </CardContent>
+      <CardContent
+        title={props.title}
+        displayLocation={props.displayLocation}
+        location={props.location}
+        linkText={props.linkText}
+        showPageLink={props.showPageLink}
+        url={props.url}
+      />
     </div>
   );
 };
 
-export default CardSingle;
+export default CardDouble;
